+++ lib/ecto/types.ex
@@ -4,3 +4,7 @@ defrecord Ecto.Interval, [:year, :month, :day, :hour, :min, :sec]
 defrecord Ecto.Binary, [:value] do
   @moduledoc false
 end
+
+defrecord Ecto.Array, [:value, :type] do
+  @moduledoc false
+end